Deflake TestFeaturesWatcher
